+++ packedForest/src/forestTypes/basicForests/rfClassification/fpForestClassificationBase.h
@@ -20,7 +20,7 @@ namespace fp {
 			std::vector<rfTree<T> > trees;
 			int numCorrect = 0;
 			int numOOB = 0;
-			std::map<std::pair<int, int>, int> pairMat;
+			std::map<std::pair<int, int>, double> pairMat;
 		public:
 
 			fpDisplayProgress printProgress;
@@ -109,7 +109,7 @@ namespace fp {
 				return bestClass;
 			}
 
-inline std::map<std::pair<int, int>, int> returnPairMat(){
+inline std::map<std::pair<int, int>, double> returnPairMat(){
                                         return pairMat;
                                 }
